# src/randomgen/openapi.yaml
# OpenAPI 3.1 contract for the RandomGen API.
#
# This file is the single source of truth for the HTTP contract (design-first):
# the service serves it verbatim at /openapi.json and renders it at /docs.
# Edit it FIRST when changing the API, then implement to match. A test pins the
# quantity limits here to the code constants so they cannot drift.
# See docs/decisions/016-design-first-openapi.md and 021-openapi-contract-version.md.
openapi: 3.1.0
info:
  title: RandomGen API
  # Contract version (the API generation), independent of the package version:
  # major = highest /api/vN served, minor = additive surface, patch = fixes (AD-21).
  # 2.1.0 adds the 429 (rate-limit) response to the generation endpoints.
  version: 2.1.0
  description: >-
    Generate random numbers from a configurable discrete distribution and score
    the sample with a Chi-Square goodness-of-fit test. `/api/v1` and `/api/v2`
    are interchangeable and differ only in the sampling algorithm.

paths:
  /api/v1/randomgen:
    get:
      summary: Generate numbers (v1)
      description: Sampling via manual inverse-CDF over `random.random()`.
      parameters:
        - $ref: '#/components/parameters/Numbers'
        - $ref: '#/components/parameters/Dist'
        - $ref: '#/components/parameters/Value'
        - $ref: '#/components/parameters/Probability'
      responses:
        '200':
          description: A generated sample and its Chi-Square quality report.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RandomGenResponse'
        '400':
          description: Invalid quantity or malformed distribution.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded; retry after the indicated interval.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v2/randomgen:
    get:
      summary: Generate numbers (v2)
      description: Sampling via `random.choices`.
      parameters:
        - $ref: '#/components/parameters/Numbers'
        - $ref: '#/components/parameters/Dist'
        - $ref: '#/components/parameters/Value'
        - $ref: '#/components/parameters/Probability'
      responses:
        '200':
          description: A generated sample and its Chi-Square quality report.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RandomGenResponse'
        '400':
          description: Invalid quantity or malformed distribution.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded; retry after the indicated interval.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /health:
    get:
      summary: Liveness check
      description: Returns 200 when the service can handle requests.
      responses:
        '200':
          description: The service is healthy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'

components:
  parameters:
    Numbers:
      name: numbers
      in: query
      required: false
      description: How many numbers to generate (1..10000). Defaults to 1000 when omitted.
      schema:
        type: integer
        default: 1000
        minimum: 1
        maximum: 10000
    Dist:
      name: dist
      in: query
      required: false
      description: >-
        Distribution as comma-separated value:probability pairs, e.g.
        `-1:0.01,0:0.3,1:0.58,2:0.1,3:0.01`. Takes precedence over the
        `value`/`probability` parameters; probabilities must sum to 1.
      schema:
        type: string
      example: '1:0.2,2:0.2,3:0.6'
    Value:
      name: value
      in: query
      required: false
      description: 'Legacy: repeated outcomes, paired by position with `probability`.'
      schema:
        type: array
        items:
          type: number
      style: form
      explode: true
    Probability:
      name: probability
      in: query
      required: false
      description: 'Legacy: repeated weights, paired by position with `value`.'
      schema:
        type: array
        items:
          type: number
      style: form
      explode: true

  schemas:
    Error:
      type: object
      description: The error envelope returned for any failed request.
      properties:
        error:
          type: string
          description: Human-readable message.
      required:
        - error
    ChiSquareTest:
      type: object
      description: Chi-Square goodness-of-fit result for the sample.
      properties:
        is_null:
          type: boolean
          description: >-
            True when the null hypothesis (the sample matches the distribution)
            is not rejected.
        chi_square:
          type: number
          description: The chi-square statistic.
        p_value:
          type: number
          description: The test p-value.
        df:
          type: integer
          description: Degrees of freedom.
    Quality:
      type: object
      description: Quality report comparing the sample to the distribution.
      properties:
        chi_square_test:
          $ref: '#/components/schemas/ChiSquareTest'
        expected_histogram:
          type: object
          description: Expected value to probability map.
          additionalProperties:
            type: number
        observed_histogram:
          type: object
          description: Observed value to frequency map for the sample.
          additionalProperties:
            type: number
    RandomGenResponse:
      type: object
      description: A generated sample and its quality report.
      properties:
        numbers:
          type: array
          description: The generated sample.
          items:
            type: number
        quality:
          $ref: '#/components/schemas/Quality'
    Health:
      type: object
      properties:
        status:
          type: string
          examples:
            - ok